feat: implement GUI for ASCII video player and add error handling messages

+++ generate_ascii_frames.py
@@ -1,5 +1,7 @@
 from PIL import Image
+from messages_gui import no_directory_frames_found, sucess_message
 import os
+
 
 def image_to_ascii(image_path, width=100):
     chars = "@%#*+=-:. "  # Gradiente de intensidade
@@ -16,11 +18,18 @@ def image_to_ascii(image_path, width=100):
     ascii_img = "\n".join([ascii_str[i:i + width] for i in range(0, len(ascii_str), width)])
     return ascii_img
 
-# Converter todos os frames
-ascii_dir = "ascii"
-os.makedirs(ascii_dir, exist_ok=True)
+def convert_frames_to_ascii(frames_dir="frames", ascii_dir="ascii", width=100):
+    # Verifica se o diretório "frames" existe
+    if not os.path.exists(frames_dir):
+        no_directory_frames_found()
+        return
 
-for file in sorted(os.listdir("frames")):
-    ascii_art = image_to_ascii(f"frames/{file}")
-    with open(f"{ascii_dir}/{file}.txt", "w") as f:
-        f.write(ascii_art)
+    # Converter todos os frames
+    os.makedirs(ascii_dir, exist_ok=True)
+
+    for file in sorted(os.listdir(frames_dir)):
+        ascii_art = image_to_ascii(f"{frames_dir}/{file}", width)
+        with open(f"{ascii_dir}/{file}.txt", "w") as f:
+            f.write(ascii_art)
+
+    sucess_message()
